GH-7919: moved folders
