+++ projects/storefrontlib/src/cms-components/product/product-images/product-images.module.ts
@@ -4,9 +4,9 @@ import { RouterModule } from '@angular/router';
 import { CmsConfig, provideDefaultConfig } from '@spartacus/core';
 import { OutletModule } from '../../../cms-structure/outlet/index';
 import { CarouselModule } from '../../../shared/components/carousel/index';
+import { ImageZoomModule } from '../../../shared/components/image-zoom/image-zoom.module';
 import { MediaModule } from '../../../shared/components/media/media.module';
 import { IconModule } from '../../misc/icon/index';
-import { ImageZoomModule } from '../image-zoom/image-zoom.module';
 import { ProductImagesComponent } from './product-images.component';
 
 @NgModule({
