+++ projects/storefrontlib/src/shared/components/image-zoom/default-image-zoom-layout.config.ts
