+++ projects/storefrontlib/src/shared/components/image-zoom/dialog/image-zoom-dialog.component.ts
