+++ projects/storefrontlib/src/shared/components/image-zoom/trigger/image-zoom-trigger.component.html
